monthlyuserreport

+++ src/main/java/com/intranet/dto/LeavesAndHolidaysDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -14,7 +15,7 @@ import java.util.List;
 @Builder
 public class LeavesAndHolidaysDTO {
 
-    private int totalLeaves;
+    private BigDecimal totalLeaves;
     private int totalHolidays;
     private List<LocalDate> leaveDates;
     private List<LocalDate> holidayDates;
